globals, bugfix, got rid of black screen on first run
bugfix is for leaving an instance group into a real group, like when queueing duo for bg blitz and the bg ends

black screen was really unnecessary. now it's just a warning to /reload that stays on top of the addon until done. temporary.lua might not be so temporary